Fix issue where MSR number in parameter where of type uint8_t instead of uint32_t.

+++ src/cpu.h
@@ -6,12 +6,12 @@
 // Read a Model-Specific-Register into dest.
 // @param msr_num: The MSR number, as found in Intel's documentation.
 // @return: The value read from the MSR.
-uint64_t read_msr(uint8_t const msr_num);
+uint64_t read_msr(uint32_t const msr_num);
 
 // Write in an MSR.
 // @param msr_num: The MSR number, as found in Intel's documentation.
 // @param val: The value to write into the MSR.
-void write_msr(uint8_t const msr_num, uint64_t const val);
+void write_msr(uint32_t const msr_num, uint64_t const val);
 
 // Read the Timestamp counter.
 // @return: The value of the TSC
@@ -24,7 +24,6 @@ uint32_t read_eflags(void);
 // Write into EFLAGS register.
 // @param eflags: The value to write into EFLAGS.
 void write_eflags(uint32_t eflags);
-
 
 // Lock up the cpu by disabling the interrupts and halting.
 void lock_up(void);
